miRPM: miRNA RPM Normalization

Functions for the analysis of count matrices and RPM normalization in miRNA studies.

Details

This package provides lightweight tools to explore, normalize, and analyze miRNA-seq count matrices using reads per million (RPM) normalization.

See also

Author

Sergio Pérez Oliveira – sergio30po@gmail.com

Manuel Fernández Sanjurjo – fernandezsanmanuel@uniovi.es

Eduardo Iglesias Gutierrez – iglesiaseduardo@uniovi.es